pyproject.toml
[tool.poetry]
name = "fake-faces"
version = "0.1.0"
description = ""
authors = ["Alex Kyllo <alex.kyllo@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.8,<3.10"
click = "^7.1.2"
tensorflow = "^2.3.1"
matplotlib = "^3.3.2"
mtcnn = "^0.1.0"
scikit-learn = "^0.23.2"
numpy = "^1.18"
eli5 = "^0.10.1"
pydot = "^1.4.1"
questionary = "^1.8.1"
python-slugify = "^4.0.1"
python-dotenv = "^0.15.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^20.8b1"
pylint = "^2.6.0"
pyclean = "^2.0.0"

[tool.poetry.scripts]
fake-faces = 'fake_faces.cli:main'

[build-system]
requires = ["poetry-core>=1.0.0a5"]
build-backend = "poetry.core.masonry.api"
